+++ components/soulmate/view/chat/ChatRule.tsx
@@ -3,7 +3,7 @@ import styled from '@emotion/styled';
 import { AnimatePresence, m } from 'framer-motion';
 import { FC, useState } from 'react';
 
-import { cardStyle } from '@/components/soulmate/view/commonStyles';
+import { cardStyle } from '@/components/soulmate/view/common/commonStyles';
 import { colors } from '@/styles/colors';
 import { MOBILE_MEDIA_QUERY } from '@/styles/mediaQuery';
 import { textStyles } from '@/styles/typography';
