+++ components/soulmate/view/chat/MissionProgress.tsx
