+++ components/soulmate/view/common/commonStyles.ts
