+++ components/soulmate/view/lobby/ExitSoulmateLink.stories.tsx
