+++ components/soulmate/view/lobby/ExitSoulmateLink.tsx
