+++ components/soulmate/view/lobby/RegisterSuccessModal.stories.tsx
