+++ components/soulmate/view/lobby/RegisterSuccessModal.tsx
